api_server_customer_management/requirements.txt
anyio==3.6.2
click==8.1.3
dnspython==2.3.0
fastapi==0.95.0
h11==0.14.0
idna==3.4
kafka-python==2.0.2
motor==3.1.1
pydantic==1.10.7
pymongo==4.3.3
sniffio==1.3.0
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1